requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.4.0a2415+gae1ec80
idna==2.10
multidict==4.7.6
python-dotenv==0.14.0
typing-extensions==3.7.4.2
websockets==8.1
yarl==1.5.0
